apply feedback - send datasource id in correct format to backend - fix dataset renaming in dataset settings - fix typo in filename
