+++ frontend/javascripts/oxalis/view/right-border-tabs/trees_tab/hide_tree_edges_icon.tsx
